refactored PlaneSearch
